requirements.txt
telethon==1.34.0
python-telegram-bot==20.0.0
flask==2.3.3
psutil==5.9.5
requests==2.31.0